Python Anywhere Deployment
The following commit encompasses all the necessary changes and new
files in order to get this working on python anywhere.
One significant change was the fact that python anywhere does not
support Celery/Redis and instead uses its own scheduled Tasks
feature. However, due to the nature of django imports, a command
had to be added to manage.py so that it could be run with all
the necessary django things.
In addition, there are many settings changes, and collectstatic
has been run as well.