Added 'sent','claimed','unclaimed','returned
The statuses of tips are now a bit more clear.
Unfortunately, all the tips prior to June will have to be discarded
as I can no longer update them and they work with the older system
.

I am unsure if the changes here resolve the issue with the blank
"sender" in "returned" tips. I'll need to wait for an unclaimed tip
to be returned to verify.