Split Month & All into Two Pages
Prior to this commit, both the data for the "All" and "Month" pages
were being loaded simultaneously along with a simple JS button
to switch back and forth between them immediately without reloading
the page.
Although it worked, the initial loading of the page was very slow.
To resolve this, the data for the month / all tips were split into
two different pages. When a user clicks the "Show All" button,
it takes a bit of time to load the page, but the overtall time of
both pages loading should be greatly improved.